requirements.txt
#LangChain Core
LangChain
langchain-core

#OpenAI Integration
langchain-openai
openai

#Anthropic Integration
langchain-anthropic

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Enviroment Variables Management
python-dotenv

#Machine Learning Utilities
numpy
scikit-learn
